Add files via upload
In this project we use H1B ban related tweets and lockdown related tweets to determine the sentiments of the people, to achieve this preprocessing was performed on the data, stop words were removed and through stemming words were reduced to their root word. The stemming is the important part of the text analytics and then the bag of words were created. This bag of words contains all the terms from which the frequently occurring term needs to be calculated ie., which term every user use in their tweets in common and after knowing the frequency count, the bar graph and word cloud is plotted which gives the clear understanding of which are the terms that frequently occurred and the information about the entire text data. We also perform topic modelling LDA where assign scores for each term and assign a topic and this is an iterative process. LDA aim is to achieve maximum likelihood and we have achieved a better model with LDA in our analysis. Then we perform the sentiment analysis on the terms and perform the analysis later we also use text clustering to the data to see how they are grouped.